Preconfiguring ofono and activating ril and udev
The primary ril daemon is now activated, the secondary depends on the dual
sim property. This is still fuzzy because lineage deactivated the auto
detection during init. We might have to have two distinct z5 builds in the
end.

Furthermore the udev rules are now available at lxc start.

boy-scout permissive mode for selinux on boot - should not hurt.

Change-Id: I950c3bc3c045f2096095e6ef6665312cf084f8c6